# apiDoc/api-docs.yml
openapi: 3.0.1
info:
  title: OpenAPI definition
  version: v0
servers:
- url: http://localhost:8080
  description: Generated server url
...
paths:
...
  /open-hours:
    get:
      tags:
      - OpenHours
      summary: Find open hours by pagination
      operationId: findOpenHours
      parameters:
      - name: page
        in: query
        description: "The specified page, start from `0`"
        required: false
        schema:
          type: integer
          format: int32
          default: 0
      - name: size
        in: query
        description: The number of records in the page
        required: false
        schema:
          type: integer
          format: int32
          default: 10
      responses:
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AppError'
        "200":
          description: Successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/OpenHours'
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AppError'
  /open-hours/findByTime:
    get:
      tags:
      - OpenHours
      summary: Find open hours by time
      operationId: findOpenHoursByTime
      parameters:
      - name: time
        in: query
        description: "The specified time (format is HH:mm), eg: 18:30"
        required: true
        schema:
          type: string
      - name: dayOfWeek
        in: query
        description: "A day of week (Sun = 0, Sat = 6)"
        required: false
        schema:
          maximum: 6
          minimum: 0
          type: integer
          format: int32
      responses:
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AppError'
        "200":
          description: Successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/OpenHours'
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AppError'
  /open-hours/findByRestaurant:
    get:
      tags:
      - OpenHours
      summary: Find open hours by restaurant ID
      operationId: findOpenHoursByRestaurant
      parameters:
      - name: restaurantId
        in: query
        description: The specified restaurant ID
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AppError'
        "200":
          description: Successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/OpenHours'
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AppError'
  /restaurants:
    get:
      tags:
      - Restaurant
      summary: Find restaurants by pagination
      operationId: findRestaurants
      parameters:
      - name: page
        in: query
        description: The specified page
        required: false
        schema:
          type: integer
          format: int32
          default: 0
      - name: size
        in: query
        description: The number of records in the page
        required: false
        schema:
          type: integer
          format: int32
          default: 10
      responses:
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AppError'
        "200":
          description: Successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Restaurant'
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AppError'
    put:
      tags:
      - Restaurant
      summary: Edit restaurant name
      operationId: updateRestaurant
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EditRestaurant'
        required: true
      responses:
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AppError'
        "200":
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Restaurant'
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AppError'
  /restaurants/findByTime:
    get:
      tags:
      - Restaurant
      summary: List all restaurants that are open at a certain time on a day of the
        week
      operationId: findRestaurantsByTime
      parameters:
      - name: time
        in: query
        description: "The specified open time (format is HH:mm), eg: 18:30"
        required: true
        schema:
          type: string
      - name: dayOfWeek
        in: query
        description: "A day of week (Sun = 0, Sat = 6)"
        required: false
        schema:
          maximum: 6
          minimum: 0
          type: integer
          format: int32
      responses:
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AppError'
        "200":
          description: Successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/RestaurantInfo'
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AppError'
  /restaurants/findByOpenHours:
    get:
      tags:
      - Restaurant
      summary: List all restaurants that are open for more or less than x hours per
        day or week
      operationId: findRestaurantsByOpenPeriod
      parameters:
      - name: openHours
        in: query
        description: "The specified open hours (min = 1, max = 24)"
        required: true
        schema:
          minimum: 1
          type: integer
          format: int32
      - name: lessThan
        in: query
        description: To indicate less or more than the specified open hours
        required: false
        schema:
          type: boolean
          default: false
      - name: perWeek
        in: query
        description: To indicate per week or per day
        required: false
        schema:
          type: boolean
          default: false
      responses:
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AppError'
        "200":
          description: Successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/RestaurantInfo'
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AppError'
  /restaurants/findByDishNumb:
    get:
      tags:
      - Restaurant
      summary: List all restaurants that have more or less than x number of dishes
        within a price range
      operationId: findRestaurantsByDishNumb
      parameters:
      - name: dishNumb
        in: query
        description: "The specified dish number (min = 1, max = 1000)"
        required: true
        schema:
          maximum: 1000
          minimum: 1
          type: integer
          format: int32
      - name: lessThan
        in: query
        description: To indicate less or more than the specified dish number
        required: false
        schema:
          type: boolean
          default: false
      - name: maxPrice
        in: query
        description: Max dish price
        required: false
        schema:
          type: number
      - name: minPrice
        in: query
        description: Min dish price
        required: false
        schema:
          type: number
      responses:
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AppError'
        "200":
          description: Successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/RestaurantInfo'
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AppError'
  /restaurants/findByName:
    get:
      tags:
      - Restaurant
      summary: Search restaurants by name
      operationId: findRestaurantsByName
      parameters:
      - name: name
        in: query
        description: The specified restaurant name
        required: true
        schema:
          type: string
      responses:
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AppError'
        "200":
          description: Successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/RestaurantInfo'
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AppError'
...
components:
  schemas:
...
    AppError:
      type: object
      properties:
        code:
          type: integer
          format: int32
        status:
          type: integer
          format: int32
        msg:
          type: string
...
    OpenHours:
      type: object
      properties:
        id:
          type: integer
          format: int64
        restaurantId:
          type: integer
          format: int64
        restaurantName:
          type: string
        dayOfWeek:
          type: integer
          format: int32
        openPeriod:
          type: integer
          format: int32
        openHour:
          type: string
        closedHour:
          type: string
    Restaurant:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        cashBalance:
          type: number
    EditRestaurant:
      required:
      - restaurantName
      type: object
      properties:
        restaurantId:
          type: integer
          format: int64
        restaurantName:
          maxLength: 128
          minLength: 0
          type: string
    RestaurantInfo:
      type: object
      properties:
        restaurantId:
          type: integer
          format: int64
        restaurantName:
          type: string
...